requirements.txt
asgiref==3.8.1
atomicwrites==1.4.1
attrs==25.1.0
certifi==2024.12.14
charset-normalizer==2.0.12
colorama==0.4.6
Django==3.2
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
flake8==5.0.4
flake8-docstrings==1.6.0
idna==3.10
iniconfig==2.0.0
mccabe==0.7.0
packaging==24.2
pluggy==0.13.1
py==1.11.0
pycodestyle==2.9.1
pydocstyle==6.3.0
pyflakes==2.5.0
PyJWT==2.1.0
pyTelegramBotAPI==4.14.1
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytest-timeout==2.1.0
python-dotenv==0.20.0
pytz==2024.2
requests==2.26.0
snowballstemmer==2.2.0
sqlparse==0.5.3
toml==0.10.2
tomli==2.2.1
typing_extensions==4.12.2
urllib3==1.26.20
